Update Soroban package versions
The versions for the Soroban packages have been updated to the latest ones. This entails several updates in different scripts, including Cargo.toml and Cargo.lock, to ensure the correct package versions are installed and used. This does not only affect the application but also the CI configuration (.github/workflows/rust.yml), where the Soroban CLI version was updated.